+++ src/generators/generators/analyze-node.generator.ts
@@ -89,14 +89,14 @@ export class AnalyzeNodeGenerator extends BaseNodeGenerator {
     const { positions: outputArrowPositions, fullCount: outputArrowCount } =
       context.arrowCalculator.calculateOutputArrowPositions(
         childResult.totalInputArrows,
-        x,
-        nodeWidth
+        childResult.fittedX,
+        childResult.fittedWidth
       );
 
     return {
-      x,
+      x: childResult.fittedX,
       y: childResult.maxChildY,
-      width: nodeWidth,
+      width: childResult.fittedWidth,
       height: nodeHeight,
       rectId,
       inputArrowCount: outputArrowCount,